Split displayable into steps files
