+++ app/models/concerns/mobile_workflow/displayable/steps/map.rb
@@ -1,0 +1,11 @@
+module MobileWorkflow
+  module Displayable
+    module Steps
+      module Map
+        def mw_map_item(id: self.id, text:, detail_text: nil, latitude:, longitude:)
+          {id: id.to_s, text: text, detailText: detail_text, latitude: latitude, longitude: longitude}.compact
+        end
+      end
+    end
+  end
+end
