+++ app/models/concerns/mobile_workflow/displayable/steps/pie_chart.rb
@@ -1,0 +1,11 @@
+module MobileWorkflow
+  module Displayable
+    module Steps
+      module PieChart
+        def mw_pie_chart_item(id: self.id, label:, value:)
+          {id: id, label: label, value: value}.compact
+        end
+      end
+    end
+  end
+end
